requirements.txt
#torch~=2.5.1+cu121
uvicorn~=0.34.0
fastapi~=0.115.8
pydantic~=2.10.6
pydantic-settings~=2.8.0
huggingface-hub~=0.29.0
pymongo~=4.11.1
requests~=2.32.3
PyYAML~=6.0.2
DBUtils~=3.1.0
openai~=1.64.0
PyPDF2~=3.0.1
langdetect~=1.0.9
transformers~=4.49.0
sentence-transformers~=3.4.1
pycryptodome~=3.21.0
python-docx~=1.1.2
python-pptx~=1.0.2
#vllm~=0.7.3
pytest~=8.3.5
RapidFuzz~=3.13.0